work toward histogram

+++ src/system/allcommonclasses/commonstructures/RawScores.java
@@ -1,5 +1,6 @@
 package system.allcommonclasses.commonstructures;
 
+import java.math.BigInteger;
 import java.util.ArrayList;
 import java.util.LinkedHashMap;
 
@@ -16,14 +17,14 @@ public class RawScores {
 
 	public ArrayList<Long> indexRankings;
 	
-	public ArrayList<Long> fieldHistogram;
+	public ArrayList<BigInteger> fieldHistogram;
 	public LinkedHashMap<String, ArrayList<Long>> variableHistograms;
 	
 	public RawScores(){
 		genuineScores = new ArrayList<Double>();
 		imposterScores = new ArrayList<Double>();
 		indexRankings = new ArrayList<Long>();
-		fieldHistogram = new ArrayList<Long>();
+		fieldHistogram = new ArrayList<BigInteger>();
 		variableHistograms = new LinkedHashMap<String, ArrayList<Long>>();
 	}
 }
